+++ Engine/Extended/NativeInternal.h
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: GPL-2.0-or-later
+#pragma once
+_Noreturn void ME_Fatal(const char *prefix, const char *message);
+const char *ME_CacheDirectory(void);
+void ME_RecordSound(void);